# diploma_work/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Structure API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/challenges/check": {
            "get": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Check and set challenges",
                "description": "Check and set challenges for the authenticated user.",
                "operationId": "0ede76bcc490d432f6124596953405d0",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "example": "application/json"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "example": "Bearer your_token"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You have completed challenges!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/challenges/store": {
            "post": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Store a new challenge",
                "description": "Store a new challenge.",
                "operationId": "8ccd4977da1f82fac2c82826582082dd",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "example": "application/json"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChallengeResource"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChallengeResource"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/challenges/index": {
            "get": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Get a list of challenges",
                "description": "Get a list of challenges.",
                "operationId": "27ee13c19c23e4d62b19c371cd600e86",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChallengeResource"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/challenges/delete/{challenge}": {
            "delete": {
                "tags": [
                    "Challenges"
                ],
                "summary": "Delete a challenge",
                "description": "Delete a challenge.",
                "operationId": "eb45b773b9b8954e31cedee74fc3dc21",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "example": "application/json"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "example": "Bearer your_token"
                    },
                    {
                        "name": "challenge",
                        "in": "path",
                        "description": "ID of the challenge to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Challenge deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Challenge not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ChallengeResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "10k Run Challenge"
                    },
                    "description": {
                        "type": "string",
                        "example": "Complete a 10km run within a month."
                    },
                    "due_type": {
                        "type": "string",
                        "example": "monthly"
                    },
                    "challenge_type": {
                        "type": "string",
                        "example": "distanceChallenge"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "points": {
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}